Rename a mismatched filename
Normally, filename is snake_cased class name.
But `HashEachMethods` class is defined in `hash_each.rb`.
It's a mismatch between filename and class name. So, I've renamed the file.
