+++ lib/rubocop.rb
@@ -165,7 +165,7 @@ require 'rubocop/cop/performance/double_start_end_with'
 require 'rubocop/cop/performance/end_with'
 require 'rubocop/cop/performance/fixed_size'
 require 'rubocop/cop/performance/flat_map'
-require 'rubocop/cop/performance/hash_each'
+require 'rubocop/cop/performance/hash_each_methods'
 require 'rubocop/cop/performance/lstrip_rstrip'
 require 'rubocop/cop/performance/range_include'
 require 'rubocop/cop/performance/redundant_block_call'
