+++ lib/rubocop/cop/performance/hash_each_methods.rb
