+++ spec/rubocop/cop/performance/hash_each_methods_spec.rb
